pyproject.toml
[tool.poetry]
name = "kws_spotter"
version = "1.1.0"
description = ""
authors = ["Jithin Sasikumar"]

[tool.poetry.dependencies]
python = "^3.8"
cffi = "1.15.1"
dataclasses = "0.6"
docker = "5.0.3"
Flask = "2.2.1"
gunicorn = "20.1.0"
h5py = "3.7.0"
hydra-core = "1.2.0"
idna = "3.3"
Jinja2 = "3.1.2"
keras = "2.9.0"
Keras-Preprocessing = "1.1.2"
keras-self-attention = "0.51.0"
librosa = "0.9.2"
Markdown = "3.4.1"
matplotlib = "3.5.2"
mlflow = "1.27.0"
numpy = "1.22.4"
oauthlib = "3.2.0"
omegaconf = "2.2.2"
pandas = "1.4.3"
Pillow = "9.2.0"
pip = "<=22.1.2"
scikit-learn = "1.1.1"
sounddevice = "0.4.4"
SoundFile = "0.10.3.post1"
tensorboard = "^2.9.1"
tensorboard-data-server = "0.6.1"
tensorboard-plugin-wit = "1.8.1"
tensorflow = "^2.9.1"
tensorflow-estimator = "^2.9.0"
tqdm = "4.64.0"
typing-extensions = "4.3.0"
urllib3 = "1.26.11"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
mypy = "^0.971"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
